requirements.txt
aiofiles==24.1.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
audioop-lts==0.2.1
beautifulsoup4==4.13.4
blinker==1.9.0
cachetools==5.5.2
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
curl_cffi==0.11.4
decorator==5.2.1
distro==1.9.0
docstring_parser==0.16
duckduckgo_search==8.1.1
fastapi==0.116.0
ffmpy==0.6.0
filelock==3.18.0
frozendict==2.4.6
fsspec==2025.5.1
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.175.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
gradio==5.35.0
gradio_client==1.10.4
greenlet==3.2.3
groovy==0.1.2
groq==0.29.0
grpcio==1.73.1
grpcio-status==1.71.2
gTTS==2.5.4
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.33.2
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
Jinja2==3.1.6
jiter==0.10.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
lxml==6.0.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
moviepy==2.2.1
mpmath==1.3.0
multitasking==0.0.11
narwhals==1.46.0
networkx==3.5
numpy==2.3.1
openai==1.93.1
orjson==3.10.18
packaging==25.0
pandas==2.3.1
peewee==3.18.1
pgvector==0.4.1
phi==0.6.7
phidata==2.7.10
pillow==11.3.0
pip==25.1.1
platformdirs==4.3.8
primp==0.15.0
proglog==0.1.12
proto-plus==1.26.1
protobuf==5.29.5
psycopg==3.2.9
psycopg-binary==3.2.9
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydantic-settings==2.10.1
pydeck==0.9.1
pydub==0.25.1
Pygments==2.19.2
pyparsing==3.2.3
pypdf==5.7.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
rich==14.0.0
rpds-py==0.26.0
rsa==4.9.1
ruff==0.12.2
safehttpx==0.1.6
safetensors==0.5.3
semantic-version==2.10.0
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.41
starlette==0.46.2
streamlit==1.46.1
sympy==1.14.0
tenacity==9.1.2
tokenizers==0.21.2
toml==0.10.2
tomli==2.2.1
tomlkit==0.13.3
torch==2.7.1
tornado==6.5.1
tqdm==4.67.1
transformers==4.53.1
typer==0.16.0
typing_extensions==4.14.1
typing-inspection==0.4.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.35.0
watchdog==6.0.0
websockets==15.0.1
yfinance==0.2.65
murf